+++ NEWS.md
@@ -1,3 +1,9 @@
+# slurmR dev
+
+*  New integer method for `get_tmp_path`, `get_job_name`, and `get_job_id`. This
+   is a fix for issue #29 (first reported by @thistleknot.)
+
+
 # slurmR 0.5-1 (CRAN)
 
 *  `sourceSlurm(plan = "collect")` won't wait forever.